+++ engine/explain.go
@@ -9,6 +9,8 @@ import (
 	"github.com/prometheus/prometheus/promql"
 
 	"github.com/thanos-io/promql-engine/execution/model"
+	"github.com/thanos-io/promql-engine/execution/telemetry"
+	"github.com/thanos-io/promql-engine/logicalplan"
 )
 
 type ExplainableQuery interface {
@@ -19,8 +21,8 @@ type ExplainableQuery interface {
 }
 
 type AnalyzeOutputNode struct {
-	OperatorTelemetry model.OperatorTelemetry `json:"telemetry,omitempty"`
-	Children          []*AnalyzeOutputNode    `json:"children,omitempty"`
+	OperatorTelemetry telemetry.OperatorTelemetry `json:"telemetry,omitempty"`
+	Children          []*AnalyzeOutputNode        `json:"children,omitempty"`
 
 	once                sync.Once
 	totalSamples        int64
@@ -62,33 +64,30 @@ func (a *AnalyzeOutputNode) aggregateSamples() {
 			childPeak := child.PeakSamples()
 			a.peakSamples = max(a.peakSamples, childPeak)
 
-			if a.OperatorTelemetry.SubQuery() {
-				// Skip aggregating from subquery to avoid double counting samples from children.
-				continue
-			}
-			if a.OperatorTelemetry.StepInvariant() {
-				// Children of step invariant operator outputs one step, but they should be counted towards all the steps.
+			switch a.OperatorTelemetry.LogicalNode().(type) {
+			case *logicalplan.Subquery:
+				// Skip aggregating samples for subquery
+			case *logicalplan.StepInvariantExpr:
 				childSamples := child.TotalSamples()
 				for i := 0; i < len(a.totalSamplesPerStep); i++ {
 					a.totalSamples += childSamples
 					a.totalSamplesPerStep[i] += childSamples
 				}
-				continue
-			}
-
-			a.totalSamples += child.TotalSamples()
-			for i, s := range child.TotalSamplesPerStep() {
-				a.totalSamplesPerStep[i] += s
+			default:
+				a.totalSamples += child.TotalSamples()
+				for i, s := range child.TotalSamplesPerStep() {
+					a.totalSamplesPerStep[i] += s
+				}
 			}
 		}
 	})
 }
 
-func analyzeQuery(obsv model.ObservableVectorOperator) *AnalyzeOutputNode {
+func analyzeQuery(obsv telemetry.ObservableVectorOperator) *AnalyzeOutputNode {
 	children := obsv.Explain()
 	var childTelemetry []*AnalyzeOutputNode
 	for _, child := range children {
-		if obsChild, ok := child.(model.ObservableVectorOperator); ok {
+		if obsChild, ok := child.(telemetry.ObservableVectorOperator); ok {
 			childTelemetry = append(childTelemetry, analyzeQuery(obsChild))
 		}
 	}
