+++ execution/function/absent.go
@@ -8,6 +8,8 @@ import (
 	"sync"
 	"time"
 
+	"github.com/thanos-io/promql-engine/execution/telemetry"
+
 	"github.com/prometheus/prometheus/model/labels"
 
 	"github.com/thanos-io/promql-engine/execution/model"
@@ -16,7 +18,7 @@ import (
 )
 
 type absentOperator struct {
-	model.OperatorTelemetry
+	telemetry.OperatorTelemetry
 
 	once     sync.Once
 	funcExpr *logicalplan.FunctionCall
@@ -36,7 +38,7 @@ func newAbsentOperator(
 		pool:     pool,
 		next:     next,
 	}
-	oper.OperatorTelemetry = model.NewTelemetry(oper, opts)
+	oper.OperatorTelemetry = telemetry.NewTelemetry(oper, opts)
 
 	return oper
 }
