+++ execution/step_invariant/step_invariant.go
@@ -8,6 +8,8 @@ import (
 	"sync"
 	"time"
 
+	"github.com/thanos-io/promql-engine/execution/telemetry"
+
 	"github.com/efficientgo/core/errors"
 	"github.com/prometheus/prometheus/model/labels"
 
@@ -31,7 +33,7 @@ type stepInvariantOperator struct {
 	step        int64
 	currentStep int64
 	stepsBatch  int
-	model.OperatorTelemetry
+	telemetry.OperatorTelemetry
 }
 
 func (u *stepInvariantOperator) Explain() (next []model.VectorOperator) {
@@ -59,7 +61,7 @@ func NewStepInvariantOperator(
 		stepsBatch:  opts.StepsBatch,
 		cacheResult: true,
 	}
-	u.OperatorTelemetry = model.NewInvariantTelemetry(u, opts)
+	u.OperatorTelemetry = telemetry.NewStepInvariantTelemetry(u, opts)
 	if u.step == 0 {
 		u.step = 1
 	}
